build: unconditionally claim -no-undefined
Libtool -no-undefined is a claim by the library designer that:
"This library, when linked, will not reference any symbols,
unless they are defined either in its own source objects or
in other explicitly listed dependencies"

In order to build a library on win32, libtool requires that the
library designer assures this.

So, either a library DOES or DOES NOT satisfy the claim: if it doesn't,
then it can't be built on windows without massive surgery. If it does,
then it does no harm to tell ALL platforms that it does so -- thus,
we unconditionally add -no-undefined to our Makefile.am/libfoo_la_LDFLAGS.